+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 #### Version 0.6.0 (not yet released)
 - AirPrint support
 - Google Cloud Print support
+- [_change_:] Renamed `isServiceAvailable` to `isAvailable`.